[TAN-8633] Share the settings panel rows and give them one row shell
